Update raw trafo for expected inputs
